requirements.txt
flask
aiohttp
bring-api
cookidoo-api
gunicorn
python-dotenv
